option to show/hide amplicon coverage
