+++ setup.py
@@ -21,7 +21,6 @@ requirements = [
     'requests',
     'pydantic',
     'markdown',
-    'tqdm'
 ]
 
 setup_requirements = ['pytest-runner', ]
