+++ wgscovplot/tmpl/wgscovplot_controlmenu_gui.html
@@ -222,7 +222,7 @@
                     <h2 class="accordion-header">
                         <button class="accordion-button collapsed" type="button" data-bs-toggle="collapse"
                                 data-bs-target="#genefeatures" aria-expanded="false" aria-controls="genefeatures">
-                            Genetic/Amplicon Features
+                            Genetic Features
                         </button>
                     </h2>
                     <div id="genefeatures" class="accordion-collapse collapse">
@@ -269,7 +269,6 @@
                     </div>
                 </div>
                 {% endif %}
-                <!---
                 {% if amplicon == True %}
                 <div class="accordion-item">
                     <h2 class="accordion-header">
@@ -281,14 +280,13 @@
                     <div id="amplicon" class="accordion-collapse collapse">
                         <div class="accordion-body">
                             <div class="form-check form-switch">
-                                <label class="form-check-label" for="toggle-genelabel">Show Amplicon Depth</label>
-                                <input class="form-check-input" type="checkbox" id="toggle-amplicon-depthlabel" unchecked>
+                                <label class="form-check-label" for="toggle-amplicon">Show Amplicon</label>
+                                <input class="form-check-input" type="checkbox" id="toggle-amplicon" checked>
                             </div>
                         </div>
                     </div>
                 </div>
                 {% endif %}
-                --->
                 <div class="accordion-item">
                     <h2 class="accordion-header">
                         <button class="accordion-button collapsed" type="button" data-bs-toggle="collapse"
